### 2016-10-18 #95 @whedon
[PRE REVIEW]: SALib: An open-source Python library for Sensitivity Analysis
**Submitting author:** jdherman (<a href="http://orcid.org/0000-0002-4081-3175">Jonathan Herman</a>)
**Repository:** <a href="https://github.com/SALib/SALib" target ="_blank">https://github.com/SALib/SALib</a>
**Version:** v1.0.0

Thanks for submitting your paper to JOSS jdherman. The JOSS editor (shown at the top of this issue) will work with you on this issue to find a reviewer for your submission before creating the main review issue.
