### 2016-10-24 #95 @kyleniemeyer
Ah great, thank you. @quaquel or @stijnvanhoey, would either of you be interested in reviewing this for JOSS? I will also be getting an existing JOSS reviewer that is not necessarily close to the relevant domain.
