### 2016-10-24 #95 @kyleniemeyer
Hi @khinsen @nirum @amoeba @jordigh would any of you be willing to review this? If so, you can respond with a 👋 (otherwise ignore). Thanks!
