### 2016-10-25 #97 @whedon
[REVIEW]: SALib: An open-source Python library for Sensitivity Analysis
**Submitting author:** jdherman (<a href="http://orcid.org/0000-0002-4081-3175">Jonathan Herman</a>)
**Repository:** <a href="https://github.com/SALib/SALib" target ="_blank">https://github.com/SALib/SALib</a>
**Version:** v1.0.0

- [x] **Authorship:** Has the submitting author (jdherman) made major contributions to the software?
