### 2016-10-25 #95 @kyleniemeyer
Alright, thanks @stijnvanhoey and @quaquel for agreeing to review! (Any of the others I pinged: you are welcome to contribute as well if you have the bandwidth.)

The review itself will take place in #97; if you haven't reviewed for JOSS before, please check out the reviewer guidelines: http://joss.theoj.org/about#reviewer_guidelines

Thanks again!
