### 2016-10-25 #97 @stijnvanhoey
Installation checked for both python 2.7 and 3.5. Able to run the minimal example in both cases.
