### 2016-10-25 #97 @quaquel
on what OS did you check it? I have acces macOS, windows 7, windows 10, and if necessary windows server 2012.
