### 2016-10-25 #97 @stijnvanhoey
Installed on Ubuntu 16.04 within conda environments.
