### 2016-10-26 4e956fe
Added all releases to changelog and authors in file
