+++ AUTHORS.rst
@@ -2,4 +2,12 @@
 Developers
 ==========
 
+* Jon Herman <jdherman8@gmail.com>
 * Will Usher <william.usher@ouce.ox.ac.uk>
+* Chris Mutel]
+* Bernardo Trindade
+* Dave Hadka
+* Matt Woodruff
+* Fernando Rios
+* Dan Hyams
+* xantares
