### 2016-10-26 976462e
Added authors and changelogs to documentation
