### 2016-10-26 81267cd
Added badges to documentation and to readme
