### 2016-10-26 c0aa843
Added citations of SALib to documentation
