+++ CITATIONS.rst
@@ -1,0 +1,34 @@
+
+=======================
+Projects that use SALib
+=======================
+
+Many projects now use the Global Sensitivity Analysis features provided by
+SALib.  Here are a selection:
+
+Software
+--------
+
+* `The City Energy Analyst <https://github.com/architecture-building-systems/CEAforArcGIS>`_
+* `pynoddy <https://github.com/flohorovicic/pynoddy>`_
+* `savvy <https://github.com/houghb/savvy>`_
+
+Blogs
+-----
+
+* `Sensitivity Analyis in Python <http://www.perrygeo.com/sensitivity-analysis-in-python.html>`_
+* `Sensitivity Analysis with SALib <http://keyboardscientist.weebly.com/blog/sensitivity-analysis-with-salib>`_
+* `Running Sobol using SALib <https://waterprogramming.wordpress.com/2013/08/05/running-sobol-sensitivity-analysis-using-salib/>`_
+* `Extensions of SALib for more complex sensitivity analyses <https://waterprogramming.wordpress.com/2014/02/11/extensions-of-salib-for-more-complex-sensitivity-analyses/>`_
+
+Videos
+------
+
+* `PyData Presentation on SALib <https://youtu.be/gkR_lz5OptU>`_
+
+
+
+If you would like to be added to this list, please submit a pull request,
+or create an issue.
+
+Many thanks for using SALib.
