### 2016-11-29 #97 @quaquel
I have successfully installed it on both mac and windows, under both python 2.7 and python 3.5.
