### 2016-11-29 #97 @quaquel
I have used the various sampling techniques successfully. The claimed functionality is thus working.
