### 2016-11-29 #97 @kyleniemeyer
Thanks @quaquel! Could you check off the remaining reviewer questions at the top? Please note that in addition to the functionality, there are questions about the documentation and paper (which @arfon shared a link to a few comments ago).

@stijnvanhoey your comments are still also welcome!
