### 2016-12-13 SALib/SALib#116 @stijnvanhoey
[JOSS review] Adaptations of the `contributing.md`
[JOSS review](https://github.com/openjournals/joss-reviews/issues/97)

The current contribution guidelines provide essential information to interpret the current implementation and objectives, such as the design principle of the decoupling of the individual methods instead of a OOP approach. This is of great help.

However, I suggest to adapt the guidelines as follows:
* The contributing guidelines do not provide additional information about how to contribute new code. For example, provide information on how to provide pull requests (maybe different for bug fixes versus new SA methods) and how revision of new methods is organised (it would be good to formalize this revision procedure, e.g. which default examples should be tested).
* Some additional information about how to report problems could be useful. This could maybe provide classification on the kind of issue: (1) pure python `Error` or (2) irregularities of the SA outcome (the functions work, but output is 'strange'). The first one is basically a bug as such, whereas the second one is more on the interpretation side. Making this difference to the user in the contribution guidelines could be helpful.
* The last section of the guidelines provides information on ideas for the future. I would suggest to get these out of the contribution guidelines and put these in separate issues, so progress can be checked for each of them and closed when finished. The latter is harder to track in the guidelines.

For these comment, I do think that the addition of some new/custom [labels](https://help.github.com/articles/creating-and-editing-labels-for-issues-and-pull-requests/) could provide sufficient abilities to support these adaptations. For example, a special label for *interpretation* questions/issues and just `bug`. The ideas of new methods could be grouped with label `addmethod`, whereas other features are `enhancement`. The usage of these labels can be explained in the guidelines.
