### 2016-12-13 SALib/SALib#117 @stijnvanhoey
[JOSS review] failed run of the tests after fresh installation
[JOSS review](https://github.com/openjournals/joss-reviews/issues/97)

When doing a fresh install of SALib in a conda py35 environment:

```
- blas=1.1=openblas
- cairo=1.14.6=0
- certifi=2016.9.26=py35_0
- cycler=0.10.0=py35_0
- fontconfig=2.11.1=5
- freetype=2.6.3=1
- gettext=0.19.7=1
- glib=2.51.0=2
- harfbuzz=1.0.6=3
- icu=56.1=4
- jpeg=9b=0
- libffi=3.2.1=2
- libgfortran=3.0.0=1
- libiconv=1.14=4
- libpng=1.6.26=0
- libtiff=4.0.6=7
- libxml2=2.9.4=3
- matplotlib=1.5.3=np111py35_3
- ncurses=5.9=10
- numpy=1.11.2=py35_blas_openblas_202
- openblas=0.2.19=0
- openssl=1.0.2j=0
- pango=1.40.3=0
- pcre=8.39=1
- pip=9.0.1=py35_1
- pixman=0.34.0=0
- pyparsing=2.1.10=py35_0
- pyqt=4.11.4=py35_4
- python=3.5.2=2
- python-dateutil=2.6.0=py35_0
- pytz=2016.10=py35_0
- qt=4.8.7=3
- readline=6.2=2
- scipy=0.18.1=np111py35_blas_openblas_201
- setuptools=29.0.1=py35_1
- sip=4.18=py35_0
- six=1.10.0=py35_1
- sqlite=3.13.0=1
- tk=8.5.19=1
- wheel=0.29.0=py35_0
- xz=5.2.2=0
- zlib=1.2.8=3
- pip:
  - salib==1.0.1
```

and running the test-functions `python setup.py test`

I get the following error report (showing the last section with the errors:
```
...
FAIL: SALib.tests.test_plotting.test_morris_covariance_plot.test
Traceback (most recent call last):
  File "/home/stijn_vanhoey/githubs/forks/SALib/.eggs/nose-1.3.7-py3.5.egg/nose/case.py", line 198, in runTest
    self.test(*self.arg)
  File "/home/stijn_vanhoey/miniconda2/envs/salib/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 53, in failer
    result = f(*args, **kwargs)
  File "/home/stijn_vanhoey/miniconda2/envs/salib/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 220, in do_test
    '(RMS %(rms).3f)'%err)
matplotlib.testing.exceptions.ImageComparisonFailure: images not close: /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot.png vs. /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot-expected.png (RMS 16.765)

FAIL: SALib.tests.test_plotting.test_morris_covariance_plot_grouped.test
Traceback (most recent call last):
  File "/home/stijn_vanhoey/githubs/forks/SALib/.eggs/nose-1.3.7-py3.5.egg/nose/case.py", line 198, in runTest
    self.test(*self.arg)
  File "/home/stijn_vanhoey/miniconda2/envs/salib/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 53, in failer
    result = f(*args, **kwargs)
  File "/home/stijn_vanhoey/miniconda2/envs/salib/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 220, in do_test
    '(RMS %(rms).3f)'%err)
matplotlib.testing.exceptions.ImageComparisonFailure: images not close: /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot_grouped.png vs. /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot_grouped-expected.png (RMS 15.536)

Ran 95 tests in 14.192s

FAILED (SKIP=6, failures=2)
```
