### 2016-12-13 SALib/SALib#118 @stijnvanhoey
[JOSS review] paper and documentation not align
[JOSS review](https://github.com/openjournals/joss-reviews/issues/97)

In the paper, the description is focusing on the 6 current implementations. However, the [consise API Reference](http://salib.readthedocs.io/en/stable/api.html) (I also checked `latest`) contains only 5 method API descriptions.
