### 2016-12-13 #97 @stijnvanhoey
In general, I would like to congratulate the authors on the current work. This library is nicely structured, well documented and easy to setup. The issues raised require minor adaptations.

Some comments:
* As no statement is made about **performance**, this is already checked as such. Still, the current contributing guidelines information is giving an advice about potential benefits by vectorizing functions, I suggest to make an issue from this, cfr. my issue https://github.com/SALib/SALib/issues/116
* With regard to the **community guidelines** and current setup of the paper, I'm wondering if the descriptions and package could not be adapted in order to formalize and simplify the potential extension with new methods (what is the workflow, what kind of tests,..). Also, the current paper is focusing on the 6 currently implemented methods, but when this would increase, this kind of information will be outdated (as well as the paper). How does this work? Should the paper be updated at that moment?
