### 2016-12-13 #97 @kyleniemeyer
Thanks @stijnvanhoey—both to you and @quaquel, what about the "Example usage" question above?

@jdherman: it looks like some issues/concerns have been raised regarding the testing, documentation, and community guidelines. Could you resolve/respond to these?
