### 2016-12-13 SALib/SALib#118 @willu47
This is due to no docstring for the fractional factorial implementation #80.
