### 2016-12-13 #97 @jdherman
Thanks @stijnvanhoey and @kyleniemeyer. I'm at a conference right now but I will resolve the issues as soon as possible (hopefully next week).
