### 2016-12-15 SALib/SALib#117 @willu47
This is a bit of a frustrating test failure!  The tests in question compare the plots made by the plotting module to a test fixture in the fixture folder. If the backend for a matplotlib is different to that used to produce the fixture, then the plot generated during the test may look rather different, but still be perfectly reasonable (just a different format).  I doubt this has anything to do with python 3.5, but rather the install or configuration of matplotlib.
