### 2016-12-15 SALib/SALib#117 @stijnvanhoey
Dear willu47, I do understand the frustration here. What I propose is that you provide me with an export from pyenv or a conda environment (`conda env export > environment.yml`) in order to reproduce your environment and rerun the tests.
