### 2016-12-15 SALib/SALib#117 @willu47
Okay, will do.  Otherwise, the tests are succeeding on the [Travis CI server](https://travis-ci.org/SALib/SALib).

Here's my `environment.yml`:

```yaml
name: salib
channels: !!python/tuple
- defaults
dependencies:
- alabaster=0.7.9=py35_0
- babel=2.3.4=py35_0
- cycler=0.10.0=py35_0
- docutils=0.12=py35_2
- freetype=2.5.5=1
- icu=54.1=0
- imagesize=0.7.1=py35_0
- jinja2=2.8=py35_1
- libpng=1.6.22=0
- markupsafe=0.23=py35_2
- matplotlib=1.5.3=np111py35_1
- mkl=11.3.3=0
- numpy=1.11.2=py35_0
- openssl=1.0.2j=0
- pip=9.0.1=py35_1
- pygments=2.1.3=py35_0
- pyparsing=2.1.4=py35_0
- pyqt=5.6.0=py35_1
- python=3.5.2=0
- python-dateutil=2.6.0=py35_0
- pytz=2016.10=py35_0
- qt=5.6.2=0
- readline=6.2=2
- requests=2.12.3=py35_0
- scipy=0.18.1=np111py35_0
- setuptools=27.2.0=py35_0
- sip=4.18=py35_0
- six=1.10.0=py35_0
- snowballstemmer=1.2.1=py35_0
- sphinx=1.5=py35_0
- sqlite=3.13.0=0
- tk=8.5.18=0
- wheel=0.29.0=py35_0
- xz=5.2.2=0
- zlib=1.2.8=3
- pip:
  - commonmark==0.5.4
  - recommonmark==0.4.0
prefix: /Users/will2/miniconda3/envs/salib
```
