### 2016-12-17 SALib/SALib#117 @stijnvanhoey
I reran the test functions with the new environment, having the same failures:

```
...
FAIL: SALib.tests.test_plotting.test_morris_covariance_plot.test
Traceback (most recent call last):
  File "/home/stijn_vanhoey/githubs/forks/SALib/.eggs/nose-1.3.7-py3.5.egg/nose/case.py", line 198, in runTest
    self.test(*self.arg)
  File "/home/stijn_vanhoey/miniconda2/envs/salibwill/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 53, in failer
    result = f(*args, **kwargs)
  File "/home/stijn_vanhoey/miniconda2/envs/salibwill/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 220, in do_test
    '(RMS %(rms).3f)'%err)
matplotlib.testing.exceptions.ImageComparisonFailure: images not close: /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot.png vs. /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot-expected.png (RMS 16.765)

FAIL: SALib.tests.test_plotting.test_morris_covariance_plot_grouped.test
Traceback (most recent call last):
  File "/home/stijn_vanhoey/githubs/forks/SALib/.eggs/nose-1.3.7-py3.5.egg/nose/case.py", line 198, in runTest
    self.test(*self.arg)
  File "/home/stijn_vanhoey/miniconda2/envs/salibwill/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 53, in failer
    result = f(*args, **kwargs)
  File "/home/stijn_vanhoey/miniconda2/envs/salibwill/lib/python3.5/site-packages/matplotlib/testing/decorators.py", line 220, in do_test
    '(RMS %(rms).3f)'%err)
matplotlib.testing.exceptions.ImageComparisonFailure: images not close: /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot_grouped.png vs. /home/stijn_vanhoey/githubs/forks/SALib/result_images/test_plotting/covariance_plot_grouped-expected.png (RMS 15.536)

Ran 95 tests in 15.109s
```

I do understand that this is rather an matplotlib issue as it is a SAlib issue. However, I do think this should be tackled as potential contributors to the SAlib will ran to these issues in the future as well. Similar issues are reported, check [this entry](http://stackoverflow.com/questions/27948126/how-can-i-write-unit-tests-against-code-that-uses-matplotlib) from the seaborn author, proposing to not use these comparison on images directly, but rather adadpt the unit tests to compare the elments inside the graph CONTRIBUTING.md(`get_xydata()`). If that would be not feasible, make sure to inform potential contributors in the [](https://github.com/SALib/SALib/blob/master/CONTRIBUTING.md) file about this comparison failure.
