### 2016-12-20 SALib/SALib#117 @jdherman
Thanks @willu47 and @stijnvanhoey. I don't think the data comparison will work here, since we're hardcoding it in the test. The point of the test is only to make sure the plotting code works. The recommendation of the link seems to be just to assume that the plot works, given the correct data -- if we make this assumption, should we remove these tests?

Another option might be to force the same backend. In `test_plotting.py` there is a commented-out line `matplotlib.use('agg')` should we uncomment that?
