### 2017-01-02 #97 @kyleniemeyer
Hi @jdherman, just wanted to ping you on these changes (happy new year!). (That said there's no deadline or anything—the paper will remain "under review" indefinitely until you make the changes.)
