### 2017-01-03 f734899
Added docstrings for fractional factorial functions
