### 2017-01-03 a0d9032
Added more information to ff docstring and added to concide api ref
