### 2017-01-03 473695f
Fixed typos and cleaned ff example
