### 2017-01-03 SALib/SALib#118 @willu47
Closed by #119
