### 2017-01-03 SALib/SALib#117 @willu47
Given that this is not critical functionality for the library, I propose we skip these tests.  I'll submit a PR shortly.
