### 2017-01-03 b23bd2c
Moved ideas for enhancements and additional methods to issue tracker
