+++ CONTRIBUTING.md
@@ -16,17 +16,6 @@ Development should follow a few objectives:
 
 * Write tests! We're using Travis CI and Coveralls. See the `SALib/tests` directory for examples. Please try to provide tests for new methods (and the existing tests can be improved, too!)
 
-Here are a few ideas for new methods/features to add:
-
-* "groups" functionality for Sobol. @willu47 added a really nice feature for the Morris method to read groups from the parameter file, and it would be great to extend this to other methods that can handle groups. For Sobol, the biggest difference would be in the cross-sampling.
-
-* "template" functionality for all methods, to put parameter samples into a model input file (like lots of older Fortran/C++ models have). Here is an example: https://gist.github.com/jdherman/fe0b81100ee83e1f6532. If you had 1000 samples, this would create 1000 copies of the model input file, with the variable names substituted with values.
-
-* Methods to consider adding: [PAWN](http://www.sciencedirect.com/science/article/pii/S1364815215000237), and [Distance-Based Generalized Sensitivity Analysis](http://link.springer.com/article/10.1007/s11004-014-9530-5)
-
-* Lots of the existing methods could be vectorized for speed. I wrote several of these before I learned the full power of numpy.
+We use Github issues to track ideas for [new methods](https://github.com/SALib/SALib/labels/add_method) and [enhancements](https://github.com/SALib/SALib/labels/enhancement).  If you are looking to contribute new methods to the library, check the labels for inspiration or to check you are not duplicating another's work.
 
 Thanks again!
-
-
-
