### 2017-01-03 a8a8c9e
Added comments on contributing code for new methods or bugs
