### 2017-01-03 3671192
Added guidance for questions and bugs
