### 2017-01-03 6c16216
small updates to CONTRIBUTING.md
