+++ CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 Admins: [Will Usher](https://github.com/willu47) and [Jon Herman](https://github.com/jdherman)
 
-Thanks to all those who contributed so far, including: [Chris Mutel](https://github.com/cmutel), [Bernardo Trindade](https://github.com/bernardoct), [Dave Hadka](https://github.com/dhadka), [Matt Woodruff](https://github.com/matthewjwoodruff), [Fernando Rios](https://github.com/zoidy), and [Dan Hyams](https://github.com/dhyams), [xantares](https://github.com/xantares). The library would be in much worse shape without these efforts.
+Thanks to all those who have [contributed so far](https://github.com/SALib/SALib/graphs/contributors)!
 
 ## Asking Questions, Reporting Bugs
 
@@ -11,37 +11,19 @@ of the library.
 
 For bugs, [create a new issue](https://github.com/SALib/SALib/issues/new)
 on Github with the label `bug`. Use this to describe the nature of the bug and
-the conditions needed to recreate it.
+the conditions needed to recreate it, including operating system and Python version.
 
 If you have a question on interpretation of results, than we may be able to help.
 Again, create an issue with your question and use the label `question_interpretation`.
 
-We cannot answer general implementation questions (such as 'how do I run my model
+We cannot answer specific implementation questions (such as 'how do I run my model
 with SALib?')
 
-You can format your questions using
-Github Markdown, which makes it easy to paste in short snippets of code
-(use backticks \` around code).  For long python error tracebacks, use a github
-[gist](https://gist.github.com/).
-
-
-## General Objectives
-
-Development should follow a few objectives:
-
-* Maintain the decoupled sample/analysis workflow, including the command-line interface. People may not have their models written in Python, so the ability to work from data files rather than Python objects is important. Sending parameter samples to the model should be left to the user.
-
-* Include only published methods. There is a separate branch that includes experimental methods, but these shouldn't be included in the master branch.
-
-* Include citations to specific implementations. There are lots of different ways to calculate Sobol indices, for example. When there isn't an obvious "right way", we'll need citations to justify the choice.
-
-* We've had some discussions about moving to an OOP setup. At this point, it would require a pretty big refactor for all of the methods. It's tough to group the methods into classes when they're decoupled, and preserving state would force people to learn a whole API. The current functional(ish) approach is easier to test, and users can do whatever they want with the numpy matrices. That's our preference for now, unless someone sees a clear way to restructure everything into classes.
-
-* Write tests! We're using Travis CI and Coveralls. See the `SALib/tests` directory for examples. Please try to provide tests for new methods (and the existing tests can be improved, too!)
+You can format your questions using Github Markdown, which makes it easy to [paste in short snippets of code](https://help.github.com/articles/creating-and-highlighting-code-blocks/). If including a very long Python error traceback, please use a Github [gist](https://gist.github.com/).
 
 ## Contributing Code
 
-If you wish to contribute some code you have written, we use pull requests to manage the process.  There are two instances in which you may contribute code.  One is to fix a bug, and the second is to add a new sensitivity analysis method, or some other new feature.
+To contribute new code, submit a pull request. There are two instances in which you may want to contribute code: to fix a bug, or to add a new feature, such as a new sensitivity analysis method.
 
 ### Fixing a Bug
 
@@ -49,22 +31,42 @@ First, create a new issue on Github with the label `bug`. Use this to describe t
 
 Then, please create a new branch with the name `bug_xxx` where xxx is the number of the issue.
 
-Then, write a test which reproduces the bug. The tests are stored in the `SALib/tests/` folder, and are run using pytest.  You can run the tests with the command `python setup.py test` from the root folder of the library.
[…]
