### 2017-01-04 SALib/SALib#116 @jdherman
Updated by #120
