### 2017-01-04 #97 @jdherman
Hi @kyleniemeyer and @stijnvanhoey -- we have updated the contributing guidelines, docs, and testing. Thanks for the suggestions, and please see what you think when you have time. Here is the pull request (now merged):
https://github.com/SALib/SALib/pull/120
