### 2017-01-04 #97 @kyleniemeyer
Thanks @jdherman!

@stijnvanhoey could you take a look at the changes, and let me know if they satisfy your outstanding issues?
