### 2017-01-06 SALib/SALib#116 @stijnvanhoey
Nice work!
