### 2017-01-06 #97 @stijnvanhoey
Thanks @jdherman and @willu47 for the work and revisions of the guidelines and documentation. I do like the scope and structure of the package. Tests and examples are available for the user as well. All issues are closed and checked, so I'm perfectly fine with publication.

Just some minor remark, the link to the advanced readme on the [github page](http://salib.github.io/SALib/) does not work: http://salib.github.io/SALib/README-advanced.md
