### 2017-01-06 #97 @willu47
Thank you @stijnvanhoey for the comments. It has been a productive (and enjoyable!) process to go through.

Regarding the broken links, I also spotted these yesterday and have made a new issue to prompt a fix. I'll get around to tidying up the page asap.

Thanks again. 👍
