### 2017-01-07 8ed8b35
Set theme jekyll-theme-cayman and migrate Page Generator content
