+++ index.md
@@ -1,12 +1,20 @@
-##Sensitivity Analysis Library (SALib)
+## Sensitivity Analysis Library (SALib)
 
 Python implementations of commonly used sensitivity analysis methods. Useful in systems modeling to calculate the effects of model inputs or exogenous factors on outputs of interest.
+
+**Documentation:** [ReadTheDocs](http://salib.readthedocs.org)
 
 **Requirements:** [NumPy](http://www.numpy.org/), [SciPy](http://www.scipy.org/), [matplotlib](http://matplotlib.org/)
 
 **Installation:** `pip install SALib` or `python setup.py install`
 
 **Build Status:** [![Build Status](https://travis-ci.org/SALib/SALib.svg?branch=master)](https://travis-ci.org/SALib/SALib)    **Test Coverage:** [![Coverage Status](https://img.shields.io/coveralls/SALib/SALib.svg)](https://coveralls.io/r/SALib/SALib)
+
+**Code Issues:** [![Code Issues](https://www.quantifiedcode.com/api/v1/project/ed62e70f899e4ec8af4ea6b2212d4b30/badge.svg)](https://www.quantifiedcode.com/app/project/ed62e70f899e4ec8af4ea6b2212d4b30)
+
+**Cite SALib:** [![DOI](https://zenodo.org/badge/15666/SALib/SALib.svg)](https://zenodo.org/badge/latestdoi/15666/SALib/SALib)
+
+ **Paper:** [![status](http://joss.theoj.org/papers/431262803744581c1d4b6a95892d3343/status.svg)](http://joss.theoj.org/papers/431262803744581c1d4b6a95892d3343)
 
 **Methods included:**
 * Sobol Sensitivity Analysis ([Sobol 2001](http://www.sciencedirect.com/science/article/pii/S0378475400002706), [Saltelli 2002](http://www.sciencedirect.com/science/article/pii/S0010465502002801), [Saltelli et al. 2010](http://www.sciencedirect.com/science/article/pii/S0010465509003087))
@@ -16,7 +24,7 @@ Python implementations of commonly used sensitivity analysis methods. Useful in 
 * Derivative-based Global Sensitivity Measure (DGSM) ([Sobol and Kucherenko 2009](http://www.sciencedirect.com/science/article/pii/S0378475409000354))
 * Fractional Factorial Sensitivity Analysis ([Saltelli et al. 2008](http://www.wiley.com/WileyCDA/WileyTitle/productCd-0470059974.html))
 
-**Contributing:** see [here](CONTRIBUTING.md)
+**Contributing:** see [here](https://github.com/SALib/SALib/blob/master/CONTRIBUTING.md)
 
 ### Quick Start
 ```python
@@ -26,11 +34,9 @@ from SALib.test_functions import Ishigami
 import numpy as np
 
 problem = {
-  'num_vars': 3, 
-  'names': ['x1', 'x2', 'x3'], 
-  'bounds': [[-3.14159265359, 3.14159265359], 
-            [-3.14159265359, 3.14159265359], 
-             [-3.14159265359, 3.14159265359]]
+  'num_vars': 3,
+  'names': ['x1', 'x2', 'x3'],
+  'bounds': [[-np.pi, np.pi]]*3
 }
 
 # Generate samples
@@ -60,9 +66,9 @@ problem = read_param_file('/path/to/file.txt')
 # ... same as above
 ```
 
-Lots of other options are included for parameter files, as well as a command-line interface. See the [advanced readme](README-advanced.md).
+Lots of other options are included for parameter files, as well as a command-line interface. See the [advanced readme](https://github.com/SALib/SALib/blob/master/README-advanced.md).
 
[…]
