### 2017-01-07 a814cd6
problem.get('groups') in analyze.morris
