### 2017-01-07 #97 @jdherman
Thanks again @kyleniemeyer and @stijnvanhoey , this has helped a lot.

I fixed the broken links on the landing page (see [here](http://salib.github.io/SALib)) as well as a few other bugs.

This is all wrapped up in a new version number, [v1.0.2](https://github.com/SALib/SALib/releases), which is also updated on PyPI.
