### 2017-01-08 #97 @kyleniemeyer
@jdherman ⚡️ Can you give a new DOI associated with this version? That's the last bit we need.
