### 2017-01-08 #97 @kyleniemeyer
Thanks @jdherman, I'll hold off closing this until the DOI is available via https://doi.org/10.5281/zenodo.233103 (it still isn't resolving for me)
