### 2017-01-08 #97 @jdherman
Ok thanks. I'm not sure how long it should take. @willu47 can you check if I did something wrong? It looks like the new DOI was created for the release:
https://zenodo.org/account/settings/github/repository/SALib/SALib

but still doesn't resolve after 4 hours. How long does it usually take?
