### 2017-01-08 #97 @kyleniemeyer
Yeah, that's a bit strange- when I upload something to Zenodo and it gives me a DOI, it generally resolves pretty much immediately, I think.

That second link is not publicly viewable, although something that has been published/released on Zenodo *will* have a URL there in addition to the DOI (what the DOI resolves to). Perhaps the release hasn't actually been published, and is just a draft?
