### 2017-01-08 #97 @jdherman
Hm still not resolving. This is a screenshot of the zenodo page, looks like it's published
![image](https://cloud.githubusercontent.com/assets/1936620/21751837/a4d9e790-d582-11e6-90dd-f45f4aeee361.png)

Maybe it just takes more time. Sorry about this!
