### 2017-01-08 #97 @arfon
> Hm still not resolving. This is a screenshot of the zenodo page, looks like it's published

My guess is that this is an issue with Zenodo. @jdherman - might be worth emailing the Zenodo folks if this isn't fixed tomorrow.
