### 2017-01-09 #97 @willu47
That's frustrating.  I have sent Zenodo support an e-mail to check up on the DOI.  It should point to [this Zenodo entry](https://zenodo.org/record/233103) but there seems to be something wrong with Zenodo as @arfon suggested.  Fingers crossed it will get fixed asap.  I have checked and rechecked @jdherman & everything seems fine.
